% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_client.R
\name{AIMessage}
\alias{AIMessage}
\title{AIMessage}
\usage{
AIMessage(response, metadata = list())
}
\description{
Class for AI messages
}
\section{Fields}{

\describe{
\item{\code{thinking}}{Character: The thinking process of the AI.}

\item{\code{content}}{Character: The content of the AI message.}
}}

\author{
EDG
}
