% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_init.R
\name{AIResponse}
\alias{AIResponse}
\title{AIResponse Class}
\usage{
AIResponse(response = list())
}
\description{
Class for AI responses that can include thinking steps, tool calls, structured output, and free
text output.
}
\section{Fields}{

\describe{
\item{\code{response}}{List of named lists including thinking, tool calls, structured output, and text output.}
}}

\author{
EDG
}
