% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Message.R
\name{as.list.Message}
\alias{as.list.Message}
\title{Convert Message to List}
\usage{
\method{as.list}{Message}(x, ...)
}
\arguments{
\item{x}{Message object}

\item{...}{Additional arguments (not used)}
}
\value{
A list representation of the Message object
}
\description{
Convert Message to List
}
\examples{
message <- Message(
  content = "What is the weather like today?",
  role = "user"
}
\author{
EDG
}
