% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_init.R
\name{as_Message}
\alias{as_Message}
\title{Convert to Message}
\usage{
as_Message(x, ...)
}
\arguments{
\item{x}{An object to convert}
}
\value{
A Message object
}
\description{
Generic method to convert various objects to rtemis.llm Message objects
}
\author{
EDG
}
