% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_init.R
\name{as_OllamaMessage}
\alias{as_OllamaMessage}
\title{Convert to OllamaMessage}
\usage{
as_OllamaMessage(x, ...)
}
\arguments{
\item{x}{An object to convert}
}
\value{
An OllamaMessage object
}
\description{
Generic method to convert various objects to kaimana OllamaMessage objects
}
\author{
EDG
}
