% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_analysis.R
\name{build_call_network}
\alias{build_call_network}
\title{Build call network (one level deep: our functions -> what they call)}
\usage{
build_call_network(r_dir = "R")
}
\arguments{
\item{r_dir}{Path to R/ directory}
}
\value{
tibble with columns: from, to, to_package
}
\description{
Build call network (one level deep: our functions -> what they call)
}
