% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_analysis.R
\name{build_frequency_table}
\alias{build_frequency_table}
\title{Build function frequency table}
\usage{
build_frequency_table(r_dir = "R")
}
\arguments{
\item{r_dir}{Path to R/ directory}
}
\value{
data.frame with columns: call, package, n_calls, n_files
}
\description{
Build function frequency table
}
