% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPT.R
\name{ChatGPT}
\alias{ChatGPT}
\title{OpenAI Chat Request Function}
\usage{
ChatGPT(
  content,
  model = "gpt-4o",
  temperature = 0.7,
  venv_name = "rpy",
  api_key = Sys.getenv("chatGPT_API_KEY"),
  base_url = "https://api.pumpkinaigc.online/v1"
)
}
\arguments{
\item{content}{A character string representing the user's question or prompt to send to the OpenAI model.}

\item{model}{A character string specifying which model to use. Default is \code{"gpt-4o"}. Other options include \code{"gpt-3.5-turbo"}, \code{"gpt-4"}, etc.}

\item{temperature}{A numeric value between 0 and 1 that controls the randomness of the model’s response. Default is \code{0.7}. Lower values make the response more deterministic.}

\item{venv_name}{A character string specifying the name of the Python virtual environment. Default is \code{"rpy"}.}

\item{api_key}{A character string representing the API key for authentication.}

\item{base_url}{A character string specifying the base URL of the OpenAI API. Default is \code{"https://api.pumpkinaigc.online/v1"}.}
}
\value{
A character string containing the model's response to the user's question.
}
\description{
This function sends a chat request to the OpenAI model and returns the model's response. It accepts various parameters to control the request, such as the model and temperature.
}
\details{
This function initializes an OpenAI client, sends the user's question to the model, and processes the response. It allows for customization of parameters like model type and temperature. The \code{tryCatch()} function ensures that errors during the API call are handled gracefully, and the request is retried if necessary.
}
