% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ollama.R
\name{check_ollama_model}
\alias{check_ollama_model}
\title{Check Ollama Model is Available}
\usage{
check_ollama_model(model)
}
\arguments{
\item{model}{Character: Name of model.}
}
\value{
Logical: TRUE if model is available, FALSE otherwise.
}
\description{
Check Ollama Model is Available
}
\author{
EDG
}
