% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_init.R
\name{check_pos_double_scalar}
\alias{check_pos_double_scalar}
\title{Check positive double scalar}
\usage{
check_pos_double_scalar(x, arg_name = deparse(substitute(x)))
}
\arguments{
\item{x}{Numeric: Value to check. Must be a single finite number strictly greater than zero.}

\item{arg_name}{Character: Argument name to use in error messages.}
}
\value{
Called for side effects. Throws an error if checks fail.
}
\description{
Check positive double scalar
}
\examples{
check_pos_double_scalar(0.001)
check_pos_double_scalar(100)
# Throw error:
try(check_pos_double_scalar(0))
try(check_pos_double_scalar(-1))
try(check_pos_double_scalar(Inf))
}
\author{
EDG
}
