% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_analysis.R
\name{classify_call}
\alias{classify_call}
\title{Classify a function call as package::func, base, or internal}
\usage{
classify_call(call_name, our_functions = character())
}
\arguments{
\item{call_name}{Character string of the function call}

\item{our_functions}{Character vector of functions defined in R/}
}
\value{
Character: the package name or "base" or "internal"
}
\description{
Classify a function call as package::func, base, or internal
}
