% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_claude.R
\name{claude_check_model}
\alias{claude_check_model}
\title{Check Claude Model Is Available}
\usage{
claude_check_model(
  x,
  base_url = CLAUDE_URL_DEFAULT,
  api_key = NULL,
  api_key_env = CLAUDE_API_KEY_ENV_DEFAULT,
  keychain_service = NULL,
  anthropic_version = CLAUDE_API_VERSION_DEFAULT
)
}
\arguments{
\item{x}{Character: Name of the model.}

\item{base_url}{Character: Base URL of the Anthropic API.}

\item{api_key}{Optional character: API key.}

\item{api_key_env}{Character: Environment variable containing the API key.}

\item{keychain_service}{Optional character: macOS Keychain service containing the API key.}

\item{anthropic_version}{Character: \code{anthropic-version} header value.}
}
\value{
NULL, invisibly, if the model is available; otherwise throws an error.
}
\description{
Check Claude Model Is Available
}
\author{
EDG
}
