% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_claude.R
\name{claude_list_models}
\alias{claude_list_models}
\title{List Claude (Anthropic) Models}
\usage{
claude_list_models(
  base_url = CLAUDE_URL_DEFAULT,
  api_key = NULL,
  api_key_env = CLAUDE_API_KEY_ENV_DEFAULT,
  keychain_service = NULL,
  anthropic_version = CLAUDE_API_VERSION_DEFAULT
)
}
\arguments{
\item{base_url}{Character: Base URL of the Anthropic API.}

\item{api_key}{Optional character: API key.}

\item{api_key_env}{Character: Environment variable containing the API key.}

\item{keychain_service}{Optional character: macOS Keychain service containing the API key.}

\item{anthropic_version}{Character: \code{anthropic-version} header value.}
}
\value{
Character vector: Model ids.
}
\description{
List Claude (Anthropic) Models
}
\author{
EDG
}
