% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm.create.HTML.visualization.R
\name{create.HTML.viz}
\alias{create.HTML.viz}
\title{Create the HTML code for Logit Leaf Model visualization}
\usage{
create.HTML.viz(LLM_OBJECT, roundingnumbers = 2)
}
\arguments{
\item{LLM_OBJECT}{output of the llm function. A LLM object containing decision rules that define segments and logistic regressions per segment}

\item{roundingnumbers}{integer stating the number of decimals in the visualization}
}
\value{
renders HTML code for a visualization
}
\description{
This function renders HTML code for a visualization of the logit leaf model.
}
