% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_init.R
\name{create_llm_message}
\alias{create_llm_message}
\title{Create agent message}
\usage{
create_llm_message(x, ...)
}
\arguments{
\item{x}{An \code{Agent} object.}

\item{content}{Character: The content of the message.}

\item{reasoning}{Optional character: The reasoning trace.}
}
\value{
An \code{LLMMessage} object.
}
\description{
Generic method to create an agent message for different backends
}
\author{
EDG
}
