% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_S7_ReAgent.R
\name{create_ReAgent}
\alias{create_ReAgent}
\title{Create a ReAgent Object}
\usage{
create_ReAgent(model_name, system_prompt)
}
\arguments{
\item{model_name}{Character: The name of the LLM model to use. Must be an Ollama model.}

\item{system_prompt}{Character: The system prompt to use.}
}
\value{
A ReAgent object.
}
\description{
Create a ReAgent Object
}
\author{
EDG
}
