% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_S7_SecondGuess.R
\name{create_SecondGuess}
\alias{create_SecondGuess}
\title{Create SecondGuess agent with Ollama}
\usage{
create_SecondGuess(
  model_name_one,
  system_prompt_one,
  temperature_one = 0.7,
  model_name_two,
  system_prompt_two = "You are a meticulous AI Assitant and your name is Kaimana.",
  temperature_two = 0.7
)
}
\arguments{
\item{model_name_one}{Character: The name of the first LLM model to use.}

\item{system_prompt_one}{Character: System prompt for the first model.}

\item{temperature_one}{Numeric: The temperature for the first model.}

\item{model_name_two}{Character: The name of the second LLM model to use.}

\item{system_prompt_two}{Character: System prompt for the second model.}

\item{temperature_two}{Numeric: The temperature for the second model.}
}
\value{
SecondGuess object
}
\description{
Create SecondGuess agent with Ollama
}
\author{
EDG
}
