% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_S7_Tool.R
\name{create_tool}
\alias{create_tool}
\title{tool}
\usage{
create_tool(name, description, parameters = list())
}
\arguments{
\item{name}{Character: The name of the tool}

\item{description}{Character: The description of the tool}

\item{parameters}{List of ToolParameter: The parameters of the tool, each  defined using
\link{tool_param}}
}
\value{
Tool object
}
\description{
Define a tool for an agent
}
\author{
EDG
}
