% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Schema.R
\docType{data}
\name{.SCHEMA_FIELD_TYPES}
\alias{.SCHEMA_FIELD_TYPES}
\title{JSON Schema Field Types}
\format{
An object of class \code{character} of length 6.
}
\usage{
.SCHEMA_FIELD_TYPES
}
\description{
JSON Schema Field Types
}
\keyword{datasets}
