% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_analysis.R
\name{extract_function_defs}
\alias{extract_function_defs}
\title{Extract function definitions from R source files}
\usage{
extract_function_defs(r_dir = "R")
}
\arguments{
\item{r_dir}{Path to R/ directory}
}
\value{
Named list: function_name -> function body (as expression)
}
\description{
Extract function definitions from R source files
}
