% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ollama.R
\name{gen_ollama}
\alias{gen_ollama}
\title{Generate Ollama Response}
\usage{
gen_ollama(model_name, prompt, output_schema = NULL, output_type = "text")
}
\arguments{
\item{prompt}{Character: Prompt to send to model.}

\item{output_schema}{List: Format of the response. Default is NULL. See Details for example.}

\item{output_type}{Character: Type of output to return: "jsonlist", "raw", "df", "text", "req"}

\item{model}{Character: Name of model to use.}
}
\value{
Character: Model response depending on \code{output_type}.
}
\description{
Generate Ollama Response
}
\details{
The \code{output_schema} argument is a list that defines the expected format of the response.
For example, a (rather generic) prompt of "Tell me about Hawaii." with output_schema:
\preformatted{
list(
  Name = list(type = "string"),
  Capital = list(type = "string"),
  Languages = list(
    type = "array",
    items = list(type = "string")
  )
)
}

would return a response like:
\preformatted{
{
  "Name": "Hawaii",
  "Capital": "Honolulu",
  "Languages": [
    "English",
    "Hawaiian"
  ]
}
}
}
\author{
EDG
}
