% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_init.R
\name{get_content}
\alias{get_content}
\title{Get content}
\usage{
get_content(x, ...)
}
\arguments{
\item{x}{An object of class AIResponse or ReasoningResponse}
}
\value{
Character if content is text, data.table if content is structured
}
\description{
Get content
}
\author{
EDG
}
