% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_datetime.R
\name{get_current_datetime}
\alias{get_current_datetime}
\title{Get Current Date and Time}
\usage{
get_current_datetime(output_type = c("json", "list"))
}
\arguments{
\item{output_type}{Character: "json" or "list".}
}
\value{
A JSON string with fields "date", "time", and "timezone".
}
\description{
Gets the current date and time in ISO format.
}
\author{
EDG
}
