% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_keychain_secret}
\alias{get_keychain_secret}
\title{Get secret from macOS Keychain}
\usage{
get_keychain_secret(service = "KAIMANA_API_KEY", account = Sys.getenv("USER"))
}
\arguments{
\item{service}{The service name in Keychain (e.g. "OPENAI_API_KEY")}

\item{account}{The account name in Keychain (e.g. "someone@gmail.com")}
}
\value{
Character string with the secret
}
\description{
Get secret from macOS Keychain
}
\details{
This function retrieves a secret (like an API key) stored in the macOS Keychain.
It uses the \code{security} command-line tool to access the Keychain.
If the key cannot be retrieved, a warning -not an error- is issued, and \code{NULL} is returned.
We avoid throwing an error to allow the calling function to handle the absence of the key.
}
