% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_init.R
\name{get_message_list}
\alias{get_message_list}
\title{Get message list}
\usage{
get_message_list(x, ...)
}
\arguments{
\item{x}{An \code{AgentState} object.}
}
\value{
A list of named lists representing messages.
}
\description{
Generic method to retrieve messages as a list of named lists for LLM APIs
}
\author{
EDG
}
