% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_init.R
\name{get_messages}
\alias{get_messages}
\title{Get messages}
\usage{
get_messages(x, ...)
}
\arguments{
\item{x}{An \code{AgentState} object.}
}
\value{
A list of \code{Message} objects.
}
\description{
Generic method to retrieve messages from \code{AgentState} objects
}
\author{
EDG
}
