% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_State.R
\name{InProcessAgentMemory}
\alias{InProcessAgentMemory}
\title{InProcessAgentMemory Class}
\usage{
InProcessAgentMemory(metadata = NULL)
}
\description{
Class for Agent state, which includes a list of Message objects and metadata.
}
\section{Fields}{

\describe{
\item{\code{state}}{Environment: Contains the list of messages in the Agent's state.}

\item{\code{metadata}}{List: The metadata associated with the Agent's state.}
}}

