% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_init.R
\name{invoke}
\alias{invoke}
\title{Invoke Method}
\usage{
invoke(x, ...)
}
\arguments{
\item{x}{An object of class LLM, ReAgent, or ReAct}

\item{...}{Additional arguments passed to specific methods}

\item{query}{Character: The query or prompt to pass to the model or agent}
}
\value{
An AIResponse object
}
\description{
Generic method for invoking LLMs and agents
}
\author{
EDG
}
