% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_S7_ReAct.R
\name{invoke.ReAct}
\alias{invoke.ReAct}
\title{Invoke ReAct Agent}
\arguments{
\item{x}{ReAct object}

\item{query}{Character: Query to pass to the agent.}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
AIResponse object
}
\description{
Invoke ReAct Agent
}
\author{
EDG
}
