% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_S7_SecondGuess.R
\name{invoke.SecondGuess}
\alias{invoke.SecondGuess}
\title{Invoke SecondGuess Agent}
\arguments{
\item{x}{SecondGuess object}

\item{query}{Character: Query to pass to the agent.}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
AIResponse object
}
\description{
Invoke SecondGuess Agent
}
\author{
EDG
}
