% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{kaimana-package}
\alias{kaimana}
\alias{kaimana-package}
\title{\pkg{kaimana}: Interface to Large Language Models}
\description{
Advanced Machine Learning made easy, efficient, reproducible
}
\section{Online Documentation and Vignettes}{

\url{https://kaimana.rtemis.org}
}

\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
