% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_client.R
\name{km_generate}
\alias{km_generate}
\title{Generate a response from Kaimana REST API}
\usage{
km_generate(
  prompt,
  output_schema = NULL,
  model_name = "granite3.3:8b",
  temperature = 0.1,
  url = "http://localhost:31650",
  verbosity = 1L
)
}
\arguments{
\item{prompt}{Character: User prompt for the API.}

\item{output_schema}{List: output schema defined using \link{make_output_schema}.}

\item{model_name}{Character: Model to use for the API.}

\item{temperature}{Numeric: Sampling temperature.}

\item{url}{Character: URL of the generate endpoint.}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
List: Parsed response from the API.
}
\description{
Generate a response from Kaimana REST API
}
\details{
This functions uses the \code{{url}/generate} and \code{{url}/generate/structured} endpoints depending on
the presence of an output schema.
}
\examples{
\dontrun{
km_generate("Hello, world!", model_name = "granite3.3:8b")
}
}
\author{
EDG
}
