% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_client.R
\name{km_get_models}
\alias{km_get_models}
\title{Get Models available in Kaimana REST API}
\usage{
km_get_models(url = "http://localhost:31650/models", verbosity = 1L)
}
\arguments{
\item{url}{Character: URL of the models endpoint.}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
Data frame: List of models available in the Kaimana API.
}
\description{
Get Models available in Kaimana REST API
}
\examples{
\dontrun{
km_get_models()
}
}
\author{
EDG
}
