% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_client.R
\name{km_health_check}
\alias{km_health_check}
\title{Run Health Check on Kaimana REST API}
\usage{
km_health_check(url = "http://localhost:31650", verbosity = 1L)
}
\arguments{
\item{url}{Character: URL of the health check endpoint.}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
List: Parsed response from the health check endpoint.
}
\description{
Run Health Check on Kaimana REST API
}
\examples{
\dontrun{
km_health_check()
}
}
\author{
EDG
}
