% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structured.R
\name{make_output_schema}
\alias{make_output_schema}
\title{Setup Output Schema}
\usage{
make_output_schema(..., required = "all")
}
\arguments{
\item{...}{Named arguments, each of which must be a list with elements "type" and "description".
"type" must be a valid JSON Schema type: "string", "number", "integer", "boolean", "array", or "object".
"description" is a string containing a short, informative description of the field.}

\item{required}{Character vector: Names of required fields or "all".}
}
\value{
A list representing the schema
}
\description{
Helper function to create a schema for structured output
}
\author{
EDG
}
