% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_format.R
\name{md_char2list}
\alias{md_char2list}
\title{Convert character vector to markdown list}
\usage{
md_char2list(x)
}
\arguments{
\item{x}{Character vector}
}
\value{
Character string formatted as markdown list
}
\description{
Convert character vector to markdown list
}
\examples{
md_char2list(c("Item 1", "Item 2", "Item 3")) |> cat()
}
\author{
EDG
}
