% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ollama.R
\name{msg_ollama}
\alias{msg_ollama}
\title{Message Ollama}
\usage{
msg_ollama(model, system = NULL, user = "Hello.", output_type = "text")
}
\arguments{
\item{model}{Character: Name of model to use.}

\item{system}{Character: Prompt to send to model (as system message).}

\item{user}{Character: Content to send to model (as user message).}

\item{output_type}{Character: Type of output to return: "df", "jsonlist", "raw", "resp", "text",
or "tools".  Default is "text".}
}
\value{
Character: Model response.
}
\description{
Message Ollama
}
\author{
SDG
}
