% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_calculator.R
\name{multiply}
\alias{multiply}
\title{Multiply two numbers}
\usage{
multiply(a, b)
}
\arguments{
\item{a}{A number}

\item{b}{A number}
}
\value{
The product of a and b
}
\description{
Multiply two numbers
}
