% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ollama.R
\name{ollama_check_model}
\alias{ollama_check_model}
\title{Check Ollama Model is Available}
\usage{
ollama_check_model(model)
}
\arguments{
\item{model}{Character: Name of model.}
}
\value{
Logical: TRUE if model is available, FALSE otherwise.
}
\description{
Check Ollama Model is Available
}
\author{
EDG
}
