% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ollama.R
\name{ollama_get_model_info}
\alias{ollama_get_model_info}
\title{Get Ollama Model Info}
\usage{
ollama_get_model_info(x = NULL, base_url = OLLAMA_URL_DEFAULT)
}
\arguments{
\item{base_url}{Character: Base URL of Ollama server.}

\item{model}{Character: Name of model. If NULL, all models are returned.}
}
\value{
List with model info if \code{x} is defined, data.table with all models' info otherwise.
}
\description{
Get Ollama Model Info
}
\author{
EDG
}
