% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ollama.R
\name{ollama_list_models}
\alias{ollama_list_models}
\title{List Ollama Models}
\usage{
ollama_list_models(base_url = OLLAMA_URL_DEFAULT)
}
\arguments{
\item{base_url}{Character: Base URL of Ollama server.}
}
\value{
Character vector: Model names.
}
\description{
List Ollama Models
}
\examples{
# Requires a running Ollama server
\dontrun{
ollama_list_models()
}
}
\author{
EDG
}
