% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_drugbank.R
\name{query_drugbank}
\alias{query_drugbank}
\title{Query DrugBank API}
\usage{
query_drugbank(query, region = "us", output = c("json", "data.table"))
}
\arguments{
\item{query}{Character: Search query.}
}
\value{
JSON or data.table with response
}
\description{
Query DrugBank API
}
\author{
EDG
}
