% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_duckduckgo.R
\name{query_duckduckgo}
\alias{query_duckduckgo}
\title{Scrape DuckDuckGo Search Results}
\usage{
query_duckduckgo(
  query,
  ia = NULL,
  return_all = FALSE,
  output = c("json", "data.table")
)
}
\arguments{
\item{query}{Character: Search query.}
}
\value{
JSON or data.table with response
}
\description{
Scrape DuckDuckGo Search Results
}
\author{
EDG
}
