% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_duckduckgo_ia.R
\name{query_duckduckgo_ia}
\alias{query_duckduckgo_ia}
\title{Query DuckDuckGo Instant Answer API}
\usage{
query_duckduckgo_ia(
  query,
  ia = NULL,
  return_all = FALSE,
  output_type = c("json", "data.table")
)
}
\arguments{
\item{query}{Character: Search query.}

\item{ia}{Character or NULL: Specific Instant Answer to request. If NULL, general search is performed.}

\item{return_all}{Logical: If TRUE, return full response as a list. If FALSE, return selected fields.}

\item{output_type}{Character: "json" or "data.table". This should be "json" when used as an agent
tool.}
}
\value{
JSON or data.table with response
}
\description{
Query DuckDuckGo Instant Answer API
}
\author{
EDG
}
