% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_semanticscholar.R
\name{query_semanticscholar}
\alias{query_semanticscholar}
\title{Query Semantic Scholar API}
\usage{
query_semanticscholar(
  query,
  fields = c("title", "year", "abstract", "url", "publicationTypes", "publicationDate",
    "openAccessPdf"),
  year = "2000-",
  limit = 5L,
  endpoint_url = "http://api.semanticscholar.org/graph/v1/paper/search",
  output_type = c("json", "data.table"),
  verbosity = 1L
)
}
\arguments{
\item{query}{Character: The search query.}

\item{fields}{Character Vector: The fields to return. See References for available fields.}

\item{year}{Character: Year filter in the format "YYYY-" or "YYYY-YYYY".}

\item{limit}{Integer: Maximum number of results to return.}

\item{endpoint_url}{Character: The Semantic Scholar API endpoint URL.}

\item{output_type}{Character: "json" or "data.table". This should be "json" when used as an agent
tool.}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
Character with JSON response or data.table
}
\description{
Query Semantic Scholar API
}
\details{
Available fields, based on https://api.semanticscholar.org/api-docs/graph#tag/Paper-Data/operation/get_graph_paper_bulk_search
Response Schema, under "data" key
\itemize{
\item{paperId}
\item{corpusId}
\item{externalIds}
\item{url}
\item{title}
\item{abstract}
\item{venue}
\item{publicationVenue}
\item{year}
\item{referenceCount}
\item{citationCount}
\item{influentialCitationCount}
\item{isOpenAccess}
\item{openAccessPdf}
\item{fieldsOfStudy}
\item{s2FieldsOfStudy}
\item{publicationTypes}
\item{publicationDate}
\item{journal}
\item{citationStyles}
\item{authors}
}
}
\author{
EDG
}
