% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_S7_ReAgent.R
\name{ReAgent}
\alias{ReAgent}
\title{ReAgent Class}
\usage{
ReAgent(model_name = character(0), system_prompt = character(0))
}
\description{
Class for ReAgent agents
}
\details{
ReAgent Class

A stateless LLM with support for reasoning.

Lather, Rinse, Repeat
}
\section{Fields}{

\describe{
\item{\code{model_name}}{Character: The name of the LLM model to use.}

\item{\code{system_prompt}}{Character: The system prompt to use.}
}}

\author{
EDG
}
