% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_S7_ReAgent.R
\name{ReasoningResponse}
\alias{ReasoningResponse}
\alias{ReAgent}
\title{ReasoningResponse Class}
\usage{
ReAgent(
  model_name = character(0),
  system_prompt = character(0),
  output_format = list()
)
}
\description{
Class for reasoning responses containing reasoning steps and final response.

Class for ReAgent agents
}
\details{
ReasoningResponse Class
}
\section{Fields}{

\describe{
\item{\code{response}}{List of format \verb{ list(role = "reasoning", content = <character>, role = "response", content = <character>)}
containing reasoning and final response.
ReAgent Class

A stateless LLM with support for reasoning.

Lather, Rinse, Repeat}

\item{\code{model_name}}{Character: The name of the LLM model to use.}

\item{\code{system_prompt}}{Character: The system prompt to use.}
}}

\author{
EDG
}
