% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_schema.R
\name{schema}
\alias{schema}
\title{Define output schema for LLM responses}
\usage{
schema(..., required = NULL, output_type = c("json", "list"))
}
\arguments{
\item{...}{Named lists defining the schema fields and their types in the form:
\code{name = list(type = "type", description = "description")}, e.g.:
\code{frequency = list(type = "numeric", description = "Frequency of the oscillator")}}

\item{required}{Optional character vector of names of required fields. If NULL, all fields are
required. Use empty character ("") for no required fields.}

\item{output_type}{Character: "json" or "list". The output format of the schema definition.}
}
\value{
Character JSON string (if output = "json") or list (if output = "list")
}
\description{
Define output schema for LLM responses
}
\details{
The \code{required} field needs to remain an array after conversion to JSON, so we use
\code{I()} to inhibit conversion to JSON object.
You can use \code{jsonlite::toJSON(your_schema, auto_unbox = TRUE, pretty = TRUE)}
to verify the output if needed. If \code{required} is a single string, it should return as
an array with one element, e.g. \verb{"required": ["frequency"]}, NOT \code{"required": "frequency"}.
}
\author{
EDG
}
