% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_calculator.R
\name{subtract}
\alias{subtract}
\title{Subtract two numbers}
\usage{
subtract(a, b)
}
\arguments{
\item{a}{A number}

\item{b}{A number}
}
\value{
The difference of a and b
}
\description{
Subtract two numbers
}
