% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_arxiv.R
\docType{data}
\name{tool_arxiv}
\alias{tool_arxiv}
\title{arXiv Search Tool}
\format{
An object of class \code{rtemis.llm::Tool} (inherits from \code{S7_object}) of length 1.
}
\usage{
tool_arxiv
}
\description{
Tool definition for arXiv search
}
\author{
EDG
}
\keyword{datasets}
