% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_datetime.R
\docType{data}
\name{tool_datetime}
\alias{tool_datetime}
\title{Get Current Date and Time}
\format{
An object of class \code{rtemis.llm::Tool} (inherits from \code{S7_object}) of length 1.
}
\usage{
tool_datetime
}
\description{
Get the current date and time in a specified format.
}
\author{
EDG
}
\keyword{datasets}
