% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_duckduckgo_ia.R
\docType{data}
\name{tool_duckduckgo_ia}
\alias{tool_duckduckgo_ia}
\title{DuckDuckGo Instant Answer Search Tool}
\format{
An object of class \code{rtemis.llm::Tool} (inherits from \code{S7_object}) of length 1.
}
\usage{
tool_duckduckgo_ia
}
\description{
Tool definition for DuckDuckGo Instant Answer search
}
\author{
EDG
}
\keyword{datasets}
