% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_S7_Tool.R
\name{tool_param}
\alias{tool_param}
\title{tool_param}
\usage{
tool_param(name, type, description, required = FALSE)
}
\arguments{
\item{name}{Character: The name of the parameter}

\item{type}{Character: The type of the parameter}

\item{description}{Character: The description of the parameter}

\item{required}{Logical: Whether the parameter is required}
}
\value{
ToolParameter object
}
\description{
Define a tool parameter schema
}
\author{
EDG
}
