% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_semanticscholar.R
\docType{data}
\name{tool_semanticscholar}
\alias{tool_semanticscholar}
\title{Semantic Scholar Search Tool}
\format{
An object of class \code{kaimana::Tool} (inherits from \code{S7_object}) of length 1.
}
\usage{
tool_semanticscholar
}
\description{
Tool definition for Semantic Scholar search
}
\author{
EDG
}
\keyword{datasets}
