% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_wikipedia.R
\docType{data}
\name{tool_wikipedia}
\alias{tool_wikipedia}
\title{Wikipedia Search Tool}
\format{
An object of class \code{kaimana::Tool} (inherits from \code{S7_object}) of length 1.
}
\usage{
tool_wikipedia
}
\description{
Tool definition for Wikipedia search
}
\author{
EDG
}
\keyword{datasets}
