% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_wikipedia.R
\name{wiki_page_text}
\alias{wiki_page_text}
\title{Get full page text of a Wikipedia page}
\usage{
wiki_page_text(pageid, language = "en", clean_refs = TRUE, verbosity = 1L)
}
\arguments{
\item{pageid}{Integer: Wikipedia page ID.}

\item{language}{Character: Language edition (default "en").}

\item{clean_refs}{Logical: Remove numeric bracketed references (default TRUE).}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
Character: Full page text (may be empty string).
}
\description{
Get full page text of a Wikipedia page
}
