% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_wikipedia.R
\name{wiki_search}
\alias{wiki_search}
\title{Search Wikipedia}
\usage{
wiki_search(query, limit = 3, language = "en", verbosity = 1L)
}
\arguments{
\item{query}{Character: Search query.}

\item{limit}{Integer: Number of results to return.}

\item{language}{Character: Language edition of Wikipedia, e.g. "en", "de".}

\item{verbosity}{Integer: Verbosity level.}

\item{clean_refs}{Logical: If TRUE, remove numeric bracketed references like \verb{[1]}, \verb{[2]} from
text.}
}
\value{
Data frame: Search results with title, snippet, and pageid.
}
\description{
Search Wikipedia
}
