% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_wikipedia.R
\name{wiki_section_text}
\alias{wiki_section_text}
\title{Get text of a specific section of a Wikipedia page}
\usage{
wiki_section_text(
  pageid,
  section_id,
  language = "en",
  clean_refs = TRUE,
  verbosity = 1L
)
}
\arguments{
\item{pageid}{Integer: Wikipedia page ID.}

\item{section_id}{Integer: Section index.}

\item{language}{Character: Language edition of Wikipedia, e.g. "en", "de".}
}
\value{
Character: Text of the specified section.
}
\description{
Get text of a specific section of a Wikipedia page
}
