% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_wikipedia.R
\name{wiki_sections}
\alias{wiki_sections}
\title{Get sections of a Wikipedia page by pageid}
\usage{
wiki_sections(pageid, language = "en", verbosity = 1L)
}
\arguments{
\item{pageid}{Integer: Wikipedia page ID.}

\item{language}{Character: Language edition of Wikipedia, e.g. "en", "de".}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
Data frame: Sections of the Wikipedia page.
}
\description{
Get sections of a Wikipedia page by pageid
}
