% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{agentapply}
\alias{agentapply}
\title{Batch submission of prompts to an Agent}
\usage{
agentapply(X, agent = NULL, extract_responses = FALSE, verbosity = 1L, ...)
}
\arguments{
\item{X}{A vector (atomic or list) to iterate over. Other objects will be coerced by
\link[base:list]{base::as.list}. Each element will form the user prompt for a separate call to the LLM agent.}

\item{agent}{Agent object to use for all calls.}

\item{extract_responses}{Logical: If \code{TRUE}, extract the content of the assistant message from
the full response. If \code{FALSE}, return the list of full Message objects.}

\item{...}{Additional arguments passed to the \code{generate} method for the agent.}
}
\value{
A list of \code{Message} objects or a list of structured responses returned by the agent
}
\description{
Batch submission of prompts to an Agent
}
\examples{
# Requires running Ollama server with the "gemma4:e4b" model
\dontrun{
  hex <- agentapply(
     c("teal", "orange", "burgundy"),
     agent = my_agent
   )
}
}
\author{
EDG
}
