% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_init.R
\name{append_message}
\alias{append_message}
\title{Append message}
\usage{
append_message(x, ...)
}
\arguments{
\item{x}{An \code{AgentState} object.}

\item{message}{A \code{Message} object to append.}
}
\value{
The updated \code{AgentState} object.
}
\description{
Generic method to append a \code{Message} object to an \code{AgentState}
}
\author{
EDG
}
