% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_init.R
\name{as_list}
\alias{as_list}
\title{Convert to R list}
\usage{
as_list(x, ...)
}
\arguments{
\item{x}{An object to convert}

\item{...}{Additional arguments for specific methods}
}
\value{
A named R list
}
\description{
Generic method to convert various objects to R lists
}
\examples{
decay_time <- field("decay_time", "Time from peak amplitude to sustain level", type = "number")
as_list(decay_time)
}
\author{
EDG
}
