% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_init.R
\name{available_tools}
\alias{available_tools}
\title{Print available tools}
\usage{
available_tools()
}
\value{
Character vector: Names of available tools
}
\description{
Print available tools
}
\author{
E.D. Gennatas
}
