% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_tools_DB.R
\name{available_tools}
\alias{available_tools}
\title{Print available tool names}
\usage{
available_tools()
}
\value{
Character vector: Names of available tools.
}
\description{
Print available tool names
}
\author{
EDG
}
