% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_init.R
\name{available_tools}
\alias{available_tools}
\title{Print built-in tools available for use by agents}
\usage{
available_tools()
}
\description{
Prints a list of built-in tool names and their descriptions
}
\author{
EDG
}
