% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_S7_LLM.R
\name{config_Ollama}
\alias{config_Ollama}
\title{Create an OllamaConfig Object}
\usage{
config_Ollama(
  model_name,
  temperature = TEMPERATURE_DEFAULT,
  base_url = OLLAMA_URL_DEFAULT
)
}
\arguments{
\item{model_name}{Character: The name of the LLM model to use. Must be an Ollama model.}

\item{temperature}{Numeric: The temperature for the model.}

\item{base_url}{Character: Base URL of Ollama server.}
}
\value{
OllamaConfig object.
}
\description{
Create an OllamaConfig Object
}
\author{
EDG
}
