% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_S7_Agent.R
\name{create_agent}
\alias{create_agent}
\title{Create an Agent}
\usage{
create_agent(
  llmconfig,
  system_prompt = SYSTEM_PROMPT_DEFAULT,
  use_memory = TRUE,
  tools = NULL,
  name = NULL
)
}
\arguments{
\item{llmconfig}{LLMConfig: The LLMConfig to use.}

\item{system_prompt}{Optional character: The system prompt to use.}

\item{use_memory}{Logical: Whether to use conversation memory.}

\item{tools}{Optional list of Tool objects: The tools available to the agent.}

\item{name}{Optional character: The name of the agent.}
}
\value{
Agent object
}
\description{
Create an Agent
}
\author{
EDG
}
