% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_S7_LLM.R
\name{create_LLM}
\alias{create_LLM}
\title{Create a LLM Object}
\usage{
create_LLM(model_name, system_prompt)
}
\arguments{
\item{model_name}{Character: The name of the LLM model to use. Must be an Ollama model.}

\item{system_prompt}{Character: The system prompt to use.}
}
\value{
A LLM object.
}
\description{
Create a LLM Object
}
\author{
EDG
}
