% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_S7_LLM.R
\name{create_Ollama}
\alias{create_Ollama}
\title{Create an Ollama Object}
\usage{
create_Ollama(
  model_name,
  system_prompt = SYSTEM_PROMPT_DEFAULT,
  temperature = TEMPERATURE_DEFAULT,
  output_schema = NULL,
  name = NULL,
  base_url = OLLAMA_URL_DEFAULT
)
}
\arguments{
\item{model_name}{Character: The name of the LLM model to use. Must be an Ollama model.}

\item{system_prompt}{Character: The system prompt to use.}

\item{temperature}{Numeric: The temperature for the model.}

\item{output_schema}{List: An optional output schema.}

\item{name}{Character or NULL: An optional name for the Ollama object.}

\item{base_url}{Character: Base URL of Ollama server.}
}
\value{
Ollama object.
}
\description{
Create an Ollama Object
}
\author{
EDG
}
