% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_S7_Tool.R
\name{create_tool}
\alias{create_tool}
\title{create_tool}
\usage{
create_tool(name, function_name, description, parameters = list())
}
\arguments{
\item{name}{Character: The name of the tool, e.g. "Wikipedia Search".}

\item{function_name}{Character: The name of the function to call, e.g. "query_wikipedia".}

\item{description}{Character: The description of the tool.}

\item{parameters}{List of \code{ToolParameter}: The parameters of the tool, each  defined using
\link{tool_param}.}
}
\value{
Tool object
}
\description{
Define a tool for an agent
}
\author{
EDG
}
