% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Schema.R
\name{field}
\alias{field}
\title{Define a schema field}
\usage{
field(
  name,
  description = name,
  type = c("string", "number", "integer", "boolean", "array", "object"),
  required = TRUE
)
}
\arguments{
\item{name}{Optional Character: The name of the field.}

\item{description}{Optional Character: A brief description of the field.}

\item{type}{Character {"string", "number", "integer", "boolean", "array", "object"}: The field
type.}

\item{required}{Logical: Whether the field is required.}
}
\value{
Field object
}
\description{
Define a schema field
}
\author{
EDG
}
