% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_schema.R
\name{field}
\alias{field}
\title{Define a schema field}
\usage{
field(
  type = c("string", "integer", "numeric", "boolean", "array", "object"),
  description
)
}
\arguments{
\item{type}{Character: "string", "integer", "numeric", "boolean", "array", "object". The field type.}

\item{description}{Character: A brief description of the field.}
}
\value{
List defining the schema field
}
\description{
Define a schema field
}
\details{
Use this function to define each individual field when calling \link{schema}.
This is a convenience function to offer type validation and allow
\code{field("numeric", "Frequency of the oscillator")} syntax.
instead of
\code{list(type = "numeric", description = "Frequency of the oscillator")}.
}
\author{
EDG
}
