% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_init.R
\name{generate}
\alias{generate}
\title{Generate Method}
\usage{
generate(x, ...)
}
\arguments{
\item{x}{An object of class LLM or Agent.}

\item{prompt}{Character: The prompt or query to pass to the model or agent.}
}
\value{
\code{Message} object or list.
}
\description{
Generic method for generating text or structured output from LLMs and Agents
}
\author{
EDG
}
