% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaimana-package.R
\docType{package}
\name{kaimana-package}
\alias{kaimana}
\alias{kaimana-package}
\title{\pkg{kaimana}: Agentic AI for the rtemis ecosystem}
\description{
The kaimana R package provides functionality for using LLMs and Agentic AI within the rtemis
ecosystem.
}
\section{Online Documentation and Vignettes}{

\url{https://kaimana.rtemis.org}
}

\author{
\strong{Maintainer}: E.D. Gennatas \email{gennatas@gmail.com} (\href{https://orcid.org/0000-0001-9280-3609}{ORCID})

}
