% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm.R
\name{llm}
\alias{llm}
\title{Create Logit Leaf Model}
\usage{
llm(df, y_value, threshold_pruning, nbr_obs_leaf)
}
\arguments{
\item{df}{dataframe containing numerical independent variables}

\item{y_value}{numerical vector of dependent variable. Currently only two class classification is supported}

\item{threshold_pruning}{threshold for impurity}

\item{nbr_obs_leaf}{the minimum number of observations in a leaf node}
}
\value{
a LLM object containing decision rules that define segments and logistic regressions per segment
}
\description{
This function creates the logit leaf model. It assumes a dataframe with numeric
values as input and a corresponding vector with dependent values.
Decision tree parameters threshold for pruning and number of observations per
leaf can be set.
}
