% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_S7_LLM.R
\name{LLM}
\alias{LLM}
\title{LLM Class}
\usage{
LLM(system_prompt = character(0))
}
\description{
Class for interacting with LLMs directly.
}
\details{
LLM Class

A stateless LLM with support for reasoning & structured output.
}
\section{Fields}{

\describe{
\item{\code{config}}{LLMConfig: The LLM configuration.}

\item{\code{system_prompt}}{Character: The system prompt to use.}
}}

\author{
EDG
}
