% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_S7_LLM.R
\name{LLM}
\alias{LLM}
\title{LLM Class}
\usage{
LLM(model_name = character(0), system_prompt = character(0))
}
\description{
Class for interacting with LLMs directly.
}
\details{
LLM Class

A stateless LLM with support for reasoning.
}
\section{Fields}{

\describe{
\item{\code{model_name}}{Character: The name of the LLM model to use.}

\item{\code{system_prompt}}{Character: The system prompt to use.}
}}

\author{
EDG
}
