% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{llmapply}
\alias{llmapply}
\title{Batch submission of prompts to an LLM}
\usage{
llmapply(
  X,
  system_prompt,
  model_name = NULL,
  backend = c("ollama", "claude", "openai"),
  temperature = 0.3,
  output_schema = NULL,
  verbosity = 1L,
  extract_responses = TRUE,
  agent = NULL,
  ...
)
}
\arguments{
\item{X}{A vector (atomic or list) to iterate over. Other objects will be coerced by
\link[base:list]{base::as.list}. Each element will form the user prompt for a separate call to the LLM agent.}

\item{system_prompt}{Character: The system prompt to use for all calls.}

\item{model_name}{Character: The name of the model to use.}

\item{backend}{Character: The LLM backend to use. One of "ollama", "claude", or "openai".}

\item{output_schema}{\code{Schema} object, named list, or JSON string defining the expected structure
of the LLM response.}

\item{verbosity}{Integer: Verbosity level for messages.}

\item{extract_responses}{Logical: If \code{TRUE}, extract the content of the assistant message from
the full response. If \code{FALSE}, return the list of full Message objects.}

\item{...}{Additional arguments passed to the \code{generate} method for the agent.}

\item{name}{Character: Name for the agent.}
}
\value{
A list of \code{Message} objects or a list of structured responses returned by the agent
}
\description{
Batch submission of prompts to an LLM
}
\examples{
# Requires running Ollama server with the "gemma4:e4b" model
\dontrun{
  hex <- llmapply(
     c("teal", "orange", "burgundy"),
     system_prompt = "Return the hexadecimal code for the following color in format #FFFFFF",
     model_name = "gemma4:e4b"
   )
}
}
\author{
EDG
}
