% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llmapply.R
\name{llmapply}
\alias{llmapply}
\title{Batch submission of prompts to an LLM agent}
\usage{
llmapply(X, agent, ...)
}
\arguments{
\item{X}{A vector (atomic or list) to iterate over. Other objects will be coerced by
\link[base:list]{base::as.list}. Each element will form the user prompt for a separate call to the LLM agent.}

\item{agent}{An object of class \code{Agent} to which the prompts will be submitted.}

\item{...}{Additional arguments passed to the \code{generate} method for the agent.}
}
\value{
A list of \code{Message} objects or a list of structured responses returned by the agent
}
\description{
Batch submission of prompts to an LLM agent
}
\author{
EDG
}
