% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_init.R
\name{map}
\alias{map}
\title{Map}
\usage{
map(x, f, ...)
}
\arguments{
\item{x}{A character vector or list to map over.}

\item{f}{An \code{LLM} or \code{Agent} object}
}
\value{
A list of \code{Message} objects (for \code{LLM}) or list of lists of \code{Message} objects
(for \code{Agent}).
}
\description{
Map
}
\details{
Use \link{responses} to retrieve just the content from the assistant messages, or \link{reasoning} to
retrieve the reasoning traces (if enabled).
}
\examples{
# Requires running Ollama server and gemma4:e4b model
\dontrun{
  llm <- create_Ollama(
    "gemma4:e4b",
    system_prompt = "Convert color to hex code using the format #FFFFFF"
  )
  x <- c("ocean teal", "california poppy orange", "bougainvillea pink")
  hex <- map(x, llm)
  hex
}
}
\author{
EDG
}
