% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ollama.R
\name{ollama_check_model}
\alias{ollama_check_model}
\title{Check Ollama Model is Available}
\usage{
ollama_check_model(x)
}
\arguments{
\item{x}{Character: Name of model.}
}
\value{
NULL, invisibly if model is available; otherwise throws an error.
}
\description{
Check Ollama Model is Available
}
\examples{
# Requires running Ollama server
\dontrun{
  ollama_check_model("gemma4:e4b")
}
}
\author{
EDG
}
