% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ollama.R
\name{ollama_get_model_info}
\alias{ollama_get_model_info}
\title{Get Ollama Model Info}
\usage{
ollama_get_model_info(model = NULL, output = "df")
}
\arguments{
\item{model}{Character: Name of model. If NULL, all models are returned.}

\item{output}{Character: Type of output to return: "df", "resp", "jsonlist", "raw", "text"}
}
\value{
Depending on \code{output}.
}
\description{
Get Ollama Model Info
}
\author{
EDG
}
