% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ollama.R
\name{ollama_get_model_info}
\alias{ollama_get_model_info}
\title{Get Ollama Model Info}
\usage{
ollama_get_model_info(x = NULL, base_url = OLLAMA_URL_DEFAULT)
}
\arguments{
\item{x}{Optional character vector: Name of model(s) to get info for. If NULL, all available
models' info is returned.}

\item{base_url}{Character: Base URL of Ollama server.}
}
\value{
data.table
}
\description{
Get Ollama Model Info
}
\examples{
# Requires a running Ollama server
\dontrun{
  ollama_get_model_info()
  ollama_get_model_info(x = "gemma4:e4b")
}
}
\author{
EDG
}
