% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_duckduckgo.R
\name{query_duckduckgo_ia}
\alias{query_duckduckgo_ia}
\title{Query DuckDuckGo Instant Answer API}
\usage{
query_duckduckgo_ia(
  query,
  ia = NULL,
  return_all = FALSE,
  output = c("json", "data.table")
)
}
\arguments{
\item{query}{Character: Search query.}
}
\value{
JSON or data.table with response
}
\description{
Query DuckDuckGo Instant Answer API
}
\author{
EDG
}
