% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtemis.llm-package.R
\docType{package}
\name{rtemis.llm-package}
\alias{rtemis.llm}
\alias{rtemis.llm-package}
\title{\pkg{rtemis.llm}: Agentic AI for the rtemis ecosystem}
\description{
rtemis.llm provides functionality to interface with Large Language Models, part of the
\href{https://www.rtemis.org}{rtemis} ecosystem. It provides an \code{Agent} class with support for
reasoning, structured output, memory management, and tool use. Allows creation of custom
LLM-based workflows and agentic AI systems using a functional user-facing frontend and an S7
backend. Includes \code{llmapply()} for quick batch LLM inference. Supports Ollama, OpenAI, and
Anthropic endpoints.
}
\section{Online Documentation and Vignettes}{

\url{https://www.rtemis.org}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://www.rtemis.org}
  \item \url{https://docs.rtemis.org/r/llm}
  \item \url{https://docs.rtemis.org/r/llm-api}
  \item Report bugs at \url{https://github.com/rtemis-org/rtemis.llm/issues}
}

}
\author{
\strong{Maintainer}: E.D. Gennatas \email{gennatas@gmail.com} (\href{https://orcid.org/0000-0001-9280-3609}{ORCID}) [copyright holder]

}
