% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtemis.llm-package.R
\docType{package}
\name{rtemis.llm-package}
\alias{rtemis.llm}
\alias{rtemis.llm-package}
\title{\pkg{rtemis.llm}: Agentic AI for the rtemis ecosystem}
\description{
The rtemis.llm R package provides functionality for using LLMs and Agentic AI within the rtemis
ecosystem.
}
\section{Online Documentation and Vignettes}{

\url{https://www.rtemis.org}
}

\author{
\strong{Maintainer}: E.D. Gennatas \email{gennatas@gmail.com} (\href{https://orcid.org/0000-0001-9280-3609}{ORCID})

}
